[Enhancement #627] Refactor StringListEdit to a more generic ValueListEdit supporting strings and numbers.
