+++ src/component/misc/ValueListEdit.scss
@@ -1,3 +1,3 @@
-.string-list-add-button {
+.value-list-add-button {
   z-index: 0 !important;
 }
